@@ -1,9 +1,5 @@
 ﻿using BBDown.Core.Entity;
-using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Text.Json;
-using System.Threading.Tasks;
 using static BBDown.Core.Entity.Entity;
 using static BBDown.Core.Util.HTTPUtil;
 
@@ -13,7 +9,7 @@ namespace BBDown.Core.Fetcher
     {
         public async Task<VInfo> FetchAsync(string id)
         {
-            id = id.Substring(3);
+            id = id[3..];
             string index = "";
             string api = $"https://api.bilibili.com/pgc/view/web/season?ep_id={id}";
             string json = await GetWebSourceAsync(api);
@@ -24,14 +20,13 @@ namespace BBDown.Core.Fetcher
             string desc = result.GetProperty("evaluate").ToString();
             string pubTime = result.GetProperty("publish").GetProperty("pub_time").ToString();
             var pages = result.GetProperty("episodes").EnumerateArray().ToList();
-            List<Page> pagesInfo = new List<Page>();
+            List<Page> pagesInfo = new();
             int i = 1;
 
             //episodes为空; 或者未包含对应epid，番外/花絮什么的
             if (pages.Count == 0 || !result.GetProperty("episodes").ToString().Contains($"/ep{id}")) 
             {
-                JsonElement sections;
-                if (result.TryGetProperty("section", out sections))
+                if (result.TryGetProperty("section", out JsonElement sections))
                 {
                     foreach (var section in sections.EnumerateArray())
                     {
@@ -48,8 +43,7 @@ namespace BBDown.Core.Fetcher
             foreach (var page in pages)
             {
                 //跳过预告
-                JsonElement badge;
-                if (page.TryGetProperty("badge", out badge) && badge.ToString() == "预告") continue;
+                if (page.TryGetProperty("badge", out JsonElement badge) && badge.ToString() == "预告") continue;
                 string res = "";
                 try
                 {
@@ -58,7 +52,7 @@ namespace BBDown.Core.Fetcher
                 catch (Exception) { }
                 string _title = page.GetProperty("title").ToString() + " " + page.GetProperty("long_title").ToString();
                 _title = _title.Trim();
-                Page p = new Page(i++,
+                Page p = new(i++,
                     page.GetProperty("aid").ToString(),
                     page.GetProperty("cid").ToString(),
                     page.GetProperty("id").ToString(),
@@ -69,15 +63,17 @@ namespace BBDown.Core.Fetcher
             }
 
 
-            var info = new VInfo();
-            info.Title = title.Trim();
-            info.Desc = desc.Trim();
-            info.Pic = cover;
-            info.PubTime = pubTime;
-            info.PagesInfo = pagesInfo;
-            info.IsBangumi = true;
-            info.IsCheese = true;
-            info.Index = index;
+            var info = new VInfo
+            {
+                Title = title.Trim(),
+                Desc = desc.Trim(),
+                Pic = cover,
+                PubTime = pubTime,
+                PagesInfo = pagesInfo,
+                IsBangumi = true,
+                IsCheese = true,
+                Index = index
+            };
 
             return info;
         }
