@@ -1,10 +1,5 @@
 ﻿using BBDown.Core.Entity;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 using System.Text.Json;
-using System.Threading.Tasks;
 using static BBDown.Core.Entity.Entity;
 using static BBDown.Core.Util.HTTPUtil;
 
@@ -14,7 +9,7 @@ namespace BBDown.Core.Fetcher
     {
         public async Task<VInfo> FetchAsync(string id)
         {
-            id = id.Substring(7);
+            id = id[7..];
             string index = "";
             string api = $"https://api.bilibili.com/pugv/view/web/season?ep_id={id}";
             string json = await GetWebSourceAsync(api);
@@ -26,10 +21,10 @@ namespace BBDown.Core.Fetcher
             string ownerName = data.GetProperty("up_info").GetProperty("uname").ToString();
             string ownerMid = data.GetProperty("up_info").GetProperty("mid").ToString();
             var pages = data.GetProperty("episodes").EnumerateArray().ToList();
-            List<Page> pagesInfo = new List<Page>();
+            List<Page> pagesInfo = new();
             foreach (var page in pages)
             {
-                Page p = new Page(page.GetProperty("index").GetInt32(),
+                Page p = new(page.GetProperty("index").GetInt32(),
                     page.GetProperty("aid").ToString(),
                     page.GetProperty("cid").ToString(),
                     page.GetProperty("id").ToString(),
@@ -44,17 +39,19 @@ namespace BBDown.Core.Fetcher
                 pagesInfo.Add(p);
             }
             string pubTime = pagesInfo.Count > 0 ? pages[0].GetProperty("release_date").ToString() : "";
-            pubTime = pubTime != "" ? (new DateTime(1970, 1, 1, 0, 0, 0, 0, DateTimeKind.Utc).AddSeconds(Convert.ToDouble(pubTime)).ToLocalTime().ToString()) : "";
+            pubTime = pubTime != "" ? new DateTime(1970, 1, 1, 0, 0, 0, 0, DateTimeKind.Utc).AddSeconds(Convert.ToDouble(pubTime)).ToLocalTime().ToString() : "";
 
-            var info = new VInfo();
-            info.Title = title.Trim();
-            info.Desc = desc.Trim();
-            info.Pic = cover;
-            info.PubTime = pubTime;
-            info.PagesInfo = pagesInfo;
-            info.IsBangumi = true;
-            info.IsCheese = true;
-            info.Index = index;
+            var info = new VInfo
+            {
+                Title = title.Trim(),
+                Desc = desc.Trim(),
+                Pic = cover,
+                PubTime = pubTime,
+                PagesInfo = pagesInfo,
+                IsBangumi = true,
+                IsCheese = true,
+                Index = index
+            };
 
             return info;
         }
