@@ -1,5 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Diagnostics;
 using System.Text;
 using System.Text.RegularExpressions;
@@ -8,7 +7,6 @@ using static BBDown.BBDownUtil;
 using static BBDown.Core.Util.SubUtil;
 using static BBDown.Core.Logger;
 using System.IO;
-using System.Linq;
 
 namespace BBDown
 {
@@ -20,7 +18,7 @@ namespace BBDown
         public static int RunExe(string app, string parms, bool customBin = false)
         {
             int code = 0;
-            Process p = new Process();
+            Process p = new();
             p.StartInfo.FileName = app;
             p.StartInfo.Arguments = parms;
             p.StartInfo.UseShellExecute = false;
@@ -44,10 +42,10 @@ namespace BBDown
             return string.IsNullOrEmpty(str) ? str : str.Replace("\"", "'");
         }
 
-        public static int MuxByMp4box(string videoPath, string audioPath, string outPath, string desc, string title, string episodeId, string pic, string lang, List<Subtitle> subs, bool audioOnly, bool videoOnly, List<ViewPoint> points)
+        public static int MuxByMp4box(string videoPath, string audioPath, string outPath, string desc, string title, string episodeId, string pic, string lang, List<Subtitle>? subs, bool audioOnly, bool videoOnly, List<ViewPoint>? points)
         {
-            StringBuilder inputArg = new StringBuilder();
-            StringBuilder metaArg = new StringBuilder();
+            StringBuilder inputArg = new();
+            StringBuilder metaArg = new();
             inputArg.Append(" -inter 500 -noprog ");
             if (!string.IsNullOrEmpty(videoPath))
                 inputArg.Append($" -add \"{videoPath}#trackID={((audioOnly && audioPath == "") ? "2" : "1")}:name=\" ");
@@ -56,7 +54,7 @@ namespace BBDown
             if (points != null && points.Count > 0)
             {
                 var meta = GetMp4boxMetaString(points);
-                var metaFile = Path.Combine(Path.GetDirectoryName(string.IsNullOrEmpty(videoPath) ? audioPath : videoPath), "chapters");
+                var metaFile = Path.Combine(Path.GetDirectoryName(string.IsNullOrEmpty(videoPath) ? audioPath : videoPath)!, "chapters");
                 File.WriteAllText(metaFile, meta);
                 inputArg.Append($" -chap  \"{metaFile}\"  ");
             }
@@ -72,7 +70,7 @@ namespace BBDown
             {
                 for (int i = 0; i < subs.Count; i++)
                 {
-                    if (File.Exists(subs[i].path) && File.ReadAllText(subs[i].path) != "")
+                    if (File.Exists(subs[i].path) && File.ReadAllText(subs[i].path!) != "")
                     {
                         inputArg.Append($" -add \"{subs[i].path}#trackID=1:name={GetSubtitleCode(subs[i].lan).Item2}:lang={GetSubtitleCode(subs[i].lan).Item1}\" ");
                     }
@@ -85,7 +83,7 @@ namespace BBDown
             return RunExe(MP4BOX, arguments, MP4BOX != "mp4box");
         }
 
-        public static int MuxAV(bool useMp4box, string videoPath, string audioPath, string outPath, string desc = "", string title = "", string episodeId = "", string pic = "", string lang = "", List<Subtitle> subs = null, bool audioOnly = false, bool videoOnly = false, List<ViewPoint> points = null)
+        public static int MuxAV(bool useMp4box, string videoPath, string audioPath, string outPath, string desc = "", string title = "", string episodeId = "", string pic = "", string lang = "", List<Subtitle>? subs = null, bool audioOnly = false, bool videoOnly = false, List<ViewPoint>? points = null)
         {
             if (audioOnly && audioPath != "") 
                 videoPath = "";
@@ -100,11 +98,11 @@ namespace BBDown
                 return MuxByMp4box(videoPath, audioPath, outPath, desc, title, episodeId, pic, lang, subs, audioOnly, videoOnly, points);
             }
 
-            if (outPath.Contains("/") && ! Directory.Exists(Path.GetDirectoryName(outPath)))
-                Directory.CreateDirectory(Path.GetDirectoryName(outPath));
+            if (outPath.Contains('/') && ! Directory.Exists(Path.GetDirectoryName(outPath)))
+                Directory.CreateDirectory(Path.GetDirectoryName(outPath)!);
             //----分析并生成-i参数
-            StringBuilder inputArg = new StringBuilder();
-            StringBuilder metaArg = new StringBuilder();
+            StringBuilder inputArg = new();
+            StringBuilder metaArg = new();
             if (!string.IsNullOrEmpty(videoPath))
                 inputArg.Append($" -i \"{videoPath}\" ");
             if (!string.IsNullOrEmpty(audioPath))
@@ -115,7 +113,7 @@ namespace BBDown
             {
                 for (int i = 0; i < subs.Count; i++)
                 {
-                    if(File.Exists(subs[i].path) && File.ReadAllText(subs[i].path) != "")
+                    if(File.Exists(subs[i].path) && File.ReadAllText(subs[i].path!) != "")
                     {
                         inputArg.Append($" -i \"{subs[i].path}\" ");
                         metaArg.Append($" -metadata:s:s:{i} handler_name=\"{GetSubtitleCode(subs[i].lan).Item2}\" -metadata:s:s:{i} language={GetSubtitleCode(subs[i].lan).Item1} ");
@@ -130,7 +128,7 @@ namespace BBDown
             if (points != null && points.Count > 0)
             {
                 var meta = GetFFmpegMetaString(points);
-                var metaFile = Path.Combine(Path.GetDirectoryName(string.IsNullOrEmpty(videoPath) ? audioPath : videoPath), "chapters");
+                var metaFile = Path.Combine(Path.GetDirectoryName(string.IsNullOrEmpty(videoPath) ? audioPath : videoPath)!, "chapters");
                 File.WriteAllText(metaFile, meta);
                 inputArg.Append($" -i \"{metaFile}\" -map_chapters {inputCount} ");
             }
@@ -164,13 +162,13 @@ namespace BBDown
             {
                 foreach (var file in files)
                 {
-                    var tmpFile = Path.Combine(Path.GetDirectoryName(file), Path.GetFileNameWithoutExtension(file) + ".ts");
+                    var tmpFile = Path.Combine(Path.GetDirectoryName(file)!, Path.GetFileNameWithoutExtension(file) + ".ts");
                     var arguments = $"-loglevel warning -y -i \"{file}\" -map 0 -c copy -f mpegts -bsf:v h264_mp4toannexb \"{tmpFile}\"";
                     LogDebug("ffmpeg命令：{0}", arguments);
                     RunExe("ffmpeg", arguments);
                     File.Delete(file);
                 }
-                var f = GetFiles(Path.GetDirectoryName(files[0]), ".ts");
+                var f = GetFiles(Path.GetDirectoryName(files[0])!, ".ts");
                 CombineMultipleFilesIntoSingleFile(f, outPath);
                 foreach (var s in f) File.Delete(s);
             }
