@@ -1,13 +1,9 @@
 ﻿using QRCoder;
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace BBDown
 {
-	public class ConsoleQRCode : AbstractQRCode
+    public class ConsoleQRCode : AbstractQRCode
 	{
 		public ConsoleQRCode() { }
 
